fix generali dopo test finale sul metallo

+++ setup-networkd/install.sh
@@ -3,9 +3,11 @@
 # verifica path di esecuzione
 SCRIPT_DIR=$(cd -- "$(dirname -- "$0")" && pwd)
 
-mv /etc/network/interfaces /etc/network/interfaces.save
+if [ -f /etc/network/interfaces ]; then
+  mv /etc/network/interfaces /etc/network/interfaces.save
+fi
 
 cp "$SCRIPT_DIR"/*.network /etc/systemd/network/
 chmod 644 /etc/systemd/network/*.network
-systemctl enable --now systemd-networkd
-systemctl disable --now systemd-networkd-wait-online
+systemctl enable systemd-networkd
+systemctl disable systemd-networkd-wait-online
